% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\docType{data}
\name{areas}
\alias{areas}
\title{Land and Water Area by FIPS}
\format{
\subsection{\code{areas}}{

A data frame in long format
\describe{
\item{fips}{
Federal Information Processing Standards code, a 5-digit code identifying
the location. The first two digits are state, last three are county. Here,
state data is given only 2 digits (e.g. Vermont is \code{50}), while counties
are given all 5 (e.g. Chittenden county is \code{50007}).
}
\item{area_sqkm}{Area of land in county or state in sq km}
\item{water_sqkm}{Area of water in county or state in sq km}
}
}
}
\usage{
areas
}
\description{
Land and water area in sq km for all US counties and states. Total area is
the sum of \code{area_sqkm} and \code{water_sqkm}.
}
\keyword{datasets}
