% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangling_functions.R
\name{calculate_census_derived}
\alias{calculate_census_derived}
\title{Calculate Census Derived Variables}
\usage{
calculate_census_derived(dat)
}
\arguments{
\item{dat}{Long format data frame}
}
\value{
Data frame with derived variables added
}
\description{
Calculate education percentages, vacancy rates, earnings ratios, disconnected youth
}
\keyword{internal}
