% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangling_functions.R
\name{calculate_census_pop_change}
\alias{calculate_census_pop_change}
\title{Calculate Census Population Change}
\usage{
calculate_census_pop_change(dat)
}
\arguments{
\item{dat}{Long format data frame with population5YearSmooth}
}
\value{
Data frame with populationChangePerc added
}
\description{
Calculate year-over-year percent change in population
}
\keyword{internal}
