% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_functions.R
\name{call_api_census_acs1}
\alias{call_api_census_acs1}
\title{Call Census ACS1 API}
\usage{
call_api_census_acs1(config)
}
\arguments{
\item{config}{Pipeline configuration list}
}
\value{
Data frame with annual population
}
\description{
Download ACS 1-year population estimates for state and county levels
}
\keyword{internal}
