% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_functions.R
\name{call_api_census_acs5}
\alias{call_api_census_acs5}
\title{Call Census ACS5 API}
\usage{
call_api_census_acs5(config)
}
\arguments{
\item{config}{Pipeline configuration list}
}
\value{
Data frame with ACS5 variables
}
\description{
Download ACS 5-year estimates for state and county levels
}
\keyword{internal}
