% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_functions.R
\name{call_api_census_voting}
\alias{call_api_census_voting}
\title{Call Census Voting API}
\usage{
call_api_census_voting(config)
}
\arguments{
\item{config}{Pipeline configuration list}
}
\value{
Data frame with voter turnout by state and year
}
\description{
Download CPS voting data for states
}
\keyword{internal}
