% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangling_functions.R
\name{create_census_metadata}
\alias{create_census_metadata}
\title{Create Census Metadata}
\usage{
create_census_metadata(dat, census_meta_path)
}
\arguments{
\item{dat}{Processed Census data}

\item{census_meta_path}{Path to metadata CSV file}
}
\value{
Metadata data frame
}
\description{
Create metadata data frame from CSV template
}
