% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parquets.R
\name{create_metadata_parquet}
\alias{create_metadata_parquet}
\title{Create Metadata Parquet}
\usage{
create_metadata_parquet(census_metadata, nass_metadata)
}
\arguments{
\item{census_data}{}

\item{nass_data}{}
}
\description{
Combined metadata into parquet file
}
