% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parquets.R
\name{create_metrics_parquet}
\alias{create_metrics_parquet}
\title{Create Metrics Parquet}
\usage{
create_metrics_parquet(census_data, nass_data)
}
\arguments{
\item{census_data}{}

\item{nass_data}{}
}
\description{
Combined metrics data into parquet file
}
