% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fips.R
\name{filter_fips}
\alias{filter_fips}
\title{Filter metrics df by FIPS code}
\usage{
filter_fips(
  df,
  scope = c("all", "counties", "new", "old", "states", "us", "neast"),
  fips_col = "fips"
)
}
\arguments{
\item{df}{A long-format metrics df.}

\item{scope}{Method by which FIPS codes are filtered. \code{all} = all states
nationally and any county in Northeast (9 + 226 = 235). \code{counties} = any county in
Northeast. This includes Connecticut's old county system and new governance
region system (226). \code{new} = all counties in Northeast, but only
Connecticut's new governance region system (218). \code{old} = all counties in
Northeast, but only Connecticut's old county system (217). \code{neast} =
Northeast states and Counties. \code{states} = Northeast states only.}

\item{fips_col}{column specifying fips code.}
}
\value{
A data.frame with filters applied.
}
\description{
Conveniently filter a long-format metrics df by FIPS code.
}
\examples{
data(metrics_example)
filter_fips(metrics_example, scope = 'all')
}
