% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_functions.R
\name{get_census_acs5_vars}
\alias{get_census_acs5_vars}
\title{Get ACS5 Variable Codes}
\usage{
get_census_acs5_vars()
}
\value{
Named list of variable codes
}
\description{
Returns named list of ACS 5-year variable codes for Census API
}
\keyword{internal}
