\name{smdata-package}
\alias{smdata-package}
\alias{smdata}
\docType{package}
\title{
Data to Accompany Smithson & Merkle, 2014
}
\description{
Data with categorical and continuous limited dependent variables, to accompany Smithson & Merkle (2014), Generalized Linear Models for Categorical and Continuous Limited Dependent Variables
}
\details{
\tabular{ll}{
Package: \tab smdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-04\cr
License: \tab GPL-2\cr
}

}
\author{
Ed Merkle & Michael Smithson

Maintainer: Ed Merkle <merklee@missouri.edu>
}
\references{
Smithson, M. & Merkle, E. C. (2014).  Generalized Linear Models for Categorical and Continuous Limited Dependent Variables.  Chapman & Hall.
}

\keyword{ package }
\keyword{ datasets }
