% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangling_functions.R
\name{smooth_census_population}
\alias{smooth_census_population}
\title{Smooth Census Population}
\usage{
smooth_census_population(dat)
}
\arguments{
\item{dat}{Long format data frame}
}
\value{
Data frame with smoothed population added
}
\description{
Fill missing years in population5Year using linear interpolation
}
\keyword{internal}
