% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_wrangling_functions.R
\name{usd_to_thousands}
\alias{usd_to_thousands}
\title{USD to thousands}
\usage{
usd_to_thousands(dat)
}
\arguments{
\item{dat}{}
}
\description{
Transform variables in USD into units of 1,000
}
