% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangling_functions.R
\name{wrangle_census_combine}
\alias{wrangle_census_combine}
\title{Combine Census API Outputs}
\usage{
wrangle_census_combine(acs5, acs1, voting)
}
\arguments{
\item{acs5}{ACS 5-year data}

\item{acs1}{ACS 1-year data}

\item{voting}{Voting data}
}
\value{
Combined data frame
}
\description{
Combine ACS5, ACS1, and voting data into single data frame
}
\keyword{internal}
